+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'stajkowski'
 license          'Apache v2.0'
 description      'LWRP Providers for API Calls'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.3'
+version          '0.2.0'
 
 supports 'ubuntu', ">= 10.04"
 supports 'debian', ">= 6.0"